split bridge into two files
