+++ tools/sim/run_bridge.py
@@ -6,8 +6,8 @@ from typing import Any
 from multiprocessing import Queue
 
 from openpilot.tools.sim.bridge.common import SimulatorBridge
-from openpilot.tools.sim.bridge.carla import CarlaBridge
-from openpilot.tools.sim.bridge.metadrive import MetaDriveBridge
+from openpilot.tools.sim.bridge.carla.carla_bridge import CarlaBridge
+from openpilot.tools.sim.bridge.metadrive.metadrive_bridge import MetaDriveBridge
 
 
 def parse_args(add_args=None):
